external/requirements.txt
orjson==3.9.9

uvicorn==0.23.2
uvloop==0.17.0

gunicorn==21.2.0

redis==5.0.1
elasticsearch[async]==8.10.1

opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.21.0

fastapi==0.103.2
pydantic==2.4.2
python-jose[cryptography]==3.3.0

pydantic-settings==2.0.3

sentry-sdk==1.42.0
